Memory-efficient representation of genomic positions

The GPos class is a container for storing a set of genomic positions (a.k.a. genomic loci). It exists in 2 flavors: UnstitchedGPos and StitchedGPos. Each flavor uses a particular internal representation:

In an UnstitchedGPos instance the positions are stored as an integer vector.

In a StitchedGPos instance the positions are stored as an IRanges object where each range represents a run of consecutive positions (i.e. a run of positions that are adjacent and in ascending order). This storage is particularly memory-efficient when the vector of positions contains long runs of consecutive positions.

Because genomic positions can be seen as genomic ranges of width 1, the GPos class extends the GenomicRanges virtual class (via the GRanges class).

## Constructor function
GPos(seqnames=NULL, pos=NULL, strand=NULL,
     ..., seqinfo=NULL, seqlengths=NULL, stitch=NA)

Arguments

seqnames, strand, ..., seqinfo, seqlengths

See documentation of the GRanges() constructor function for a description of these arguments.

pos

NULL, or an integer or numeric vector, or an IRanges or IPos object (or other IntegerRanges derivative). If not NULL, GPos() will try to turn it into an IPos derivative with IPos(pos, stitch=stitch).

When pos is an IRanges object (or other IntegerRanges derivative), each range in it is interpreted as a run of consecutive positions.

stitch

TRUE, FALSE, or NA (the default).

Controls which internal representation should be used: StitchedGPos (when stitch is TRUE) or UnstitchedGPos (when stitch is FALSE).

When stitch is NA (the default), which internal representation will be used depends on the flavour of the IPos derivative returned by IPos(pos): UnstitchedGPos if IPos(pos) returns an UnstitchedIPos instance, and StitchedGPos if it returns a StitchedIPos instance.

Details

Even though a GRanges object can be used for storing genomic positions, using a GPos object is more efficient. In particular the memory footprint of an UnstitchedGPos object is typically about half that of a GRanges object.

OTOH the memory footprint of a StitchedGPos object can vary a lot but will never be worse than that of a GRanges object. However it will reduce dramatically if the vector of positions contains long runs of consecutive positions. In the worst case scenario (i.e. when the object contains no consecutive positions) its memory footprint will be the same as that of a GRanges object.

Like for any Vector derivative, the length of a GPos object cannot exceed .Machine$integer.max (i.e. 2^31 on most platforms). GPos() will return an error if pos contains too many positions.

Value

An UnstitchedGPos or StitchedGPos object.

Accessors

Getters

GPos objects support the same set of getters as other GenomicRanges derivatives (i.e. seqnames(), ranges(), start(), end(), strand(), mcols(), seqinfo(), etc...), plus the pos() getter which is equivalent to start() or end(). See ?GenomicRanges for the list of getters supported by GenomicRanges derivatives.

Note that ranges() returns an IPos derivative instead of the IRanges object that one gets with other GenomicRanges derivatives. To get an IRanges object, you need to call ranges() again on this IPos derivative i.e. do ranges(ranges(x)) on GPos object x.

Setters

Like GRanges objects, GPos derivatives support the following setters:

The seqnames() and strand() setters.

The names(), mcols() and metadata() setters.

The family of setters that operate on the seqinfo component of an object: seqlevels(), seqlevelsStyle(), seqlengths(), isCircular(), genome(), and seqinfo(). These setters are defined and documented in the GenomeInfoDb package.

However, there is no pos() setter for GPos derivatives at the moment (although one might be added in the future).

Coercion

From UnstitchedGPos to StitchedGPos and vice-versa: coercion back and forth between UnstitchedGPos and StitchedGPos is supported via as(x, "StitchedGPos") and as(x, "UnstitchedGPos"). This is the most efficient and recommended way to switch between the 2 internal representations. Note that this switch can have dramatic consequences on memory usage so is for advanced users only. End users should almost never need to do this switch when following a typical workflow.

From GenomicRanges to UnstitchedGPos, StitchedGPos, or GPos: A GenomicRanges derivative x in which all the ranges have a width of 1 can be coerced to an UnstitchedGPos or StitchedGPos object with as(x, "UnstitchedGPos") or as(x, "StitchedGPos"), respectively. For convenience as(x, "GPos") is supported and is equivalent to as(x, "UnstitchedGPos").

From GPos to GRanges: A GPos derivative x can be coerced to a GRanges object with as(x, "GRanges"). However be aware that the resulting object can use thousands times (or more) memory than x! See "MEMORY USAGE" in the Examples section below.

From GPos to ordinary R objects: Like with any other GenomicRanges derivative, as.character(), as.factor(), and as.data.frame() work on a GPos derivative x. Note however that as.data.frame(x) returns a data frame with a pos column (containing pos(x)) instead of the start, end, and width columns that one gets with other GenomicRanges derivatives.

Subsetting

A GPos derivative can be subsetted exactly like a GRanges object.

Concatenation

GPos derivatives can be concatenated with c() or append(). See ?c in the S4Vectors package for more information about concatenating Vector derivatives.

Splitting and Relisting

Like with any other GRanges object, split() and relist() work on a GPos derivative.

Note

Internal representation of GPos objects has changed in GenomicRanges 1.29.10 (Bioc 3.6). Update any old object x with: x <- updateObject(x, verbose=TRUE).

Author

Hervé Pagès; based on ideas borrowed from Georg Stricker georg.stricker@in.tum.de and Julien Gagneur gagneur@in.tum.de

See also

Examples

showClass("GPos")  # shows the known subclasses
#> Virtual Class "GPos" [package "GenomicRanges"]
#> 
#> Slots:
#>                                                                       
#> Name:           ranges elementMetadata     elementType        metadata
#> Class:            IPos       DataFrame       character            list
#>                                                       
#> Name:         seqnames          strand         seqinfo
#> Class:             Rle             Rle         Seqinfo
#> 
#> Extends: 
#> Class "GenomicPos", directly
#> Class "GRanges", directly
#> Class "GenomicRanges", by class "GenomicPos", distance 2
#> Class "Pos", by class "GenomicPos", distance 2
#> Class "Ranges", by class "GenomicPos", distance 3
#> Class "GenomicRanges_OR_missing", by class "GenomicPos", distance 3
#> Class "List", by class "GenomicPos", distance 4
#> Class "Vector", by class "GenomicPos", distance 5
#> Class "list_OR_List", by class "GenomicPos", distance 5
#> Class "Annotated", by class "GenomicPos", distance 6
#> Class "vector_OR_Vector", by class "GenomicPos", distance 6
#> Class "GenomicRanges_OR_GenomicRangesList", by class "GenomicRanges", distance 3
#> Class "GenomicRanges_OR_GRangesList", by class "GenomicRanges", distance 3
#> 
#> Known Subclasses: "UnstitchedGPos", "StitchedGPos"

## ---------------------------------------------------------------------
## BASIC EXAMPLES
## ---------------------------------------------------------------------

## Example 1:
gpos1a <- GPos(seqnames=Rle(c("chr1", "chr2", "chr1"), c(10, 6, 4)),
               pos=c(44:53, 5:10, 2:5))
gpos1a  # unstitched
#> UnstitchedGPos object with 20 positions and 0 metadata columns:
#>        seqnames       pos strand
#>           <Rle> <integer>  <Rle>
#>    [1]     chr1        44      *
#>    [2]     chr1        45      *
#>    [3]     chr1        46      *
#>    [4]     chr1        47      *
#>    [5]     chr1        48      *
#>    ...      ...       ...    ...
#>   [16]     chr2        10      *
#>   [17]     chr1         2      *
#>   [18]     chr1         3      *
#>   [19]     chr1         4      *
#>   [20]     chr1         5      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

length(gpos1a)
#> [1] 20
seqnames(gpos1a)
#> factor-Rle of length 20 with 3 runs
#>   Lengths:   10    6    4
#>   Values : chr1 chr2 chr1
#> Levels(2): chr1 chr2
pos(gpos1a)  # same as 'start(gpos1a)' and 'end(gpos1a)'
#>  [1] 44 45 46 47 48 49 50 51 52 53  5  6  7  8  9 10  2  3  4  5
strand(gpos1a)
#> factor-Rle of length 20 with 1 run
#>   Lengths: 20
#>   Values :  *
#> Levels(3): + - *
as.character(gpos1a)
#>  [1] "chr1:44" "chr1:45" "chr1:46" "chr1:47" "chr1:48" "chr1:49" "chr1:50"
#>  [8] "chr1:51" "chr1:52" "chr1:53" "chr2:5"  "chr2:6"  "chr2:7"  "chr2:8" 
#> [15] "chr2:9"  "chr2:10" "chr1:2"  "chr1:3"  "chr1:4"  "chr1:5" 
as.data.frame(gpos1a)
#>    seqnames pos strand
#> 1      chr1  44      *
#> 2      chr1  45      *
#> 3      chr1  46      *
#> 4      chr1  47      *
#> 5      chr1  48      *
#> 6      chr1  49      *
#> 7      chr1  50      *
#> 8      chr1  51      *
#> 9      chr1  52      *
#> 10     chr1  53      *
#> 11     chr2   5      *
#> 12     chr2   6      *
#> 13     chr2   7      *
#> 14     chr2   8      *
#> 15     chr2   9      *
#> 16     chr2  10      *
#> 17     chr1   2      *
#> 18     chr1   3      *
#> 19     chr1   4      *
#> 20     chr1   5      *
as(gpos1a, "GRanges")
#> GRanges object with 20 ranges and 0 metadata columns:
#>        seqnames    ranges strand
#>           <Rle> <IRanges>  <Rle>
#>    [1]     chr1        44      *
#>    [2]     chr1        45      *
#>    [3]     chr1        46      *
#>    [4]     chr1        47      *
#>    [5]     chr1        48      *
#>    ...      ...       ...    ...
#>   [16]     chr2        10      *
#>   [17]     chr1         2      *
#>   [18]     chr1         3      *
#>   [19]     chr1         4      *
#>   [20]     chr1         5      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
as.data.frame(as(gpos1a, "GRanges"))
#>    seqnames start end width strand
#> 1      chr1    44  44     1      *
#> 2      chr1    45  45     1      *
#> 3      chr1    46  46     1      *
#> 4      chr1    47  47     1      *
#> 5      chr1    48  48     1      *
#> 6      chr1    49  49     1      *
#> 7      chr1    50  50     1      *
#> 8      chr1    51  51     1      *
#> 9      chr1    52  52     1      *
#> 10     chr1    53  53     1      *
#> 11     chr2     5   5     1      *
#> 12     chr2     6   6     1      *
#> 13     chr2     7   7     1      *
#> 14     chr2     8   8     1      *
#> 15     chr2     9   9     1      *
#> 16     chr2    10  10     1      *
#> 17     chr1     2   2     1      *
#> 18     chr1     3   3     1      *
#> 19     chr1     4   4     1      *
#> 20     chr1     5   5     1      *
gpos1a[9:17]
#> UnstitchedGPos object with 9 positions and 0 metadata columns:
#>       seqnames       pos strand
#>          <Rle> <integer>  <Rle>
#>   [1]     chr1        52      *
#>   [2]     chr1        53      *
#>   [3]     chr2         5      *
#>   [4]     chr2         6      *
#>   [5]     chr2         7      *
#>   [6]     chr2         8      *
#>   [7]     chr2         9      *
#>   [8]     chr2        10      *
#>   [9]     chr1         2      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

gpos1b <- GPos(seqnames=Rle(c("chr1", "chr2", "chr1"), c(10, 6, 4)),
               pos=c(44:53, 5:10, 2:5), stitch=TRUE)
gpos1b  # stitched
#> StitchedGPos object with 20 positions and 0 metadata columns:
#>        seqnames       pos strand
#>           <Rle> <integer>  <Rle>
#>    [1]     chr1        44      *
#>    [2]     chr1        45      *
#>    [3]     chr1        46      *
#>    [4]     chr1        47      *
#>    [5]     chr1        48      *
#>    ...      ...       ...    ...
#>   [16]     chr2        10      *
#>   [17]     chr1         2      *
#>   [18]     chr1         3      *
#>   [19]     chr1         4      *
#>   [20]     chr1         5      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

## 'gpos1a' and 'gpos1b' are semantically equivalent, only their
## internal representations differ:
all(gpos1a == gpos1b)
#> [1] TRUE

gpos1c <- GPos(c("chr1:44-53", "chr2:5-10", "chr1:2-5"))
gpos1c  # stitched
#> StitchedGPos object with 20 positions and 0 metadata columns:
#>        seqnames       pos strand
#>           <Rle> <integer>  <Rle>
#>    [1]     chr1        44      *
#>    [2]     chr1        45      *
#>    [3]     chr1        46      *
#>    [4]     chr1        47      *
#>    [5]     chr1        48      *
#>    ...      ...       ...    ...
#>   [16]     chr2        10      *
#>   [17]     chr1         2      *
#>   [18]     chr1         3      *
#>   [19]     chr1         4      *
#>   [20]     chr1         5      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

identical(gpos1b, gpos1c)
#> [1] TRUE

## Example 2:
pos_runs <- GRanges("chrI", IRanges(c(1, 6, 12, 17), c(5, 10, 16, 20)),
                    strand=c("*", "-", "-", "+"))
gpos2 <- GPos(pos_runs)
gpos2  # stitched
#> StitchedGPos object with 19 positions and 0 metadata columns:
#>        seqnames       pos strand
#>           <Rle> <integer>  <Rle>
#>    [1]     chrI         1      *
#>    [2]     chrI         2      *
#>    [3]     chrI         3      *
#>    [4]     chrI         4      *
#>    [5]     chrI         5      *
#>    ...      ...       ...    ...
#>   [15]     chrI        16      -
#>   [16]     chrI        17      +
#>   [17]     chrI        18      +
#>   [18]     chrI        19      +
#>   [19]     chrI        20      +
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
strand(gpos2)
#> factor-Rle of length 19 with 3 runs
#>   Lengths:  5 10  4
#>   Values :  *  -  +
#> Levels(3): + - *

## Example 3:
gpos3A <- gpos3B <- GPos(c("chrI:1-1000", "chrI:1005-2000"))
npos <- length(gpos3A)

mcols(gpos3A)$sample <- Rle("sA")
sA_counts <- sample(10, npos, replace=TRUE)
mcols(gpos3A)$counts <- sA_counts

mcols(gpos3B)$sample <- Rle("sB")
sB_counts <- sample(10, npos, replace=TRUE)
mcols(gpos3B)$counts <- sB_counts

gpos3 <- c(gpos3A, gpos3B)
gpos3
#> StitchedGPos object with 3992 positions and 2 metadata columns:
#>          seqnames       pos strand | sample    counts
#>             <Rle> <integer>  <Rle> |  <Rle> <integer>
#>      [1]     chrI         1      * |     sA         9
#>      [2]     chrI         2      * |     sA         9
#>      [3]     chrI         3      * |     sA        10
#>      [4]     chrI         4      * |     sA        10
#>      [5]     chrI         5      * |     sA         7
#>      ...      ...       ...    ... .    ...       ...
#>   [3988]     chrI      1996      * |     sB         7
#>   [3989]     chrI      1997      * |     sB         7
#>   [3990]     chrI      1998      * |     sB         1
#>   [3991]     chrI      1999      * |     sB         6
#>   [3992]     chrI      2000      * |     sB         8
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Example 4:
library(BSgenome.Scerevisiae.UCSC.sacCer2)
#> Loading required package: BSgenome
#> Loading required package: Biostrings
#> Loading required package: XVector
#> 
#> Attaching package: ‘Biostrings’
#> The following object is masked from ‘package:base’:
#> 
#>     strsplit
#> Loading required package: rtracklayer
genome <- BSgenome.Scerevisiae.UCSC.sacCer2
gpos4 <- GPos(seqinfo(genome))
gpos4  # all the positions along the genome are represented
#> StitchedGPos object with 12162995 positions and 0 metadata columns:
#>              seqnames       pos strand
#>                 <Rle> <integer>  <Rle>
#>          [1]     chrI         1      *
#>          [2]     chrI         2      *
#>          [3]     chrI         3      *
#>          [4]     chrI         4      *
#>          [5]     chrI         5      *
#>          ...      ...       ...    ...
#>   [12162991]  2micron      6314      *
#>   [12162992]  2micron      6315      *
#>   [12162993]  2micron      6316      *
#>   [12162994]  2micron      6317      *
#>   [12162995]  2micron      6318      *
#>   -------
#>   seqinfo: 18 sequences (2 circular) from sacCer2 genome
mcols(gpos4)$dna <- do.call("c", unname(as.list(genome)))
gpos4
#> StitchedGPos object with 12162995 positions and 1 metadata column:
#>              seqnames       pos strand |         dna
#>                 <Rle> <integer>  <Rle> | <DNAString>
#>          [1]     chrI         1      * |           C
#>          [2]     chrI         2      * |           C
#>          [3]     chrI         3      * |           A
#>          [4]     chrI         4      * |           C
#>          [5]     chrI         5      * |           A
#>          ...      ...       ...    ... .         ...
#>   [12162991]  2micron      6314      * |           A
#>   [12162992]  2micron      6315      * |           A
#>   [12162993]  2micron      6316      * |           C
#>   [12162994]  2micron      6317      * |           G
#>   [12162995]  2micron      6318      * |           A
#>   -------
#>   seqinfo: 18 sequences (2 circular) from sacCer2 genome

## Note however that, like for any Vector derivative, the length of a
## GPos derivative cannot exceed '.Machine$integer.max' (i.e. 2^31 on
## most platforms) so the above only works with a "small" genome.
## For example it doesn't work with the Human genome:
library(TxDb.Hsapiens.UCSC.hg19.knownGene)
#> Loading required package: GenomicFeatures
#> Loading required package: AnnotationDbi
#> Loading required package: Biobase
#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.
if (FALSE) {
  GPos(seqinfo(TxDb.Hsapiens.UCSC.hg19.knownGene))  # error!
}

## You can use isSmallGenome() to check upfront whether the genome is
## "small" or not.
isSmallGenome(genome)  # TRUE
#> [1] TRUE
isSmallGenome(TxDb.Hsapiens.UCSC.hg19.knownGene)  # FALSE
#> [1] FALSE

## ---------------------------------------------------------------------
## MEMORY USAGE
## ---------------------------------------------------------------------

## Coercion to GRanges works...
gr4 <- as(gpos4, "GRanges")
gr4
#> GRanges object with 12162995 ranges and 1 metadata column:
#>              seqnames    ranges strand |         dna
#>                 <Rle> <IRanges>  <Rle> | <DNAString>
#>          [1]     chrI         1      * |           C
#>          [2]     chrI         2      * |           C
#>          [3]     chrI         3      * |           A
#>          [4]     chrI         4      * |           C
#>          [5]     chrI         5      * |           A
#>          ...      ...       ...    ... .         ...
#>   [12162991]  2micron      6314      * |           A
#>   [12162992]  2micron      6315      * |           A
#>   [12162993]  2micron      6316      * |           C
#>   [12162994]  2micron      6317      * |           G
#>   [12162995]  2micron      6318      * |           A
#>   -------
#>   seqinfo: 18 sequences (2 circular) from sacCer2 genome
## ... but is generally not a good idea:
object.size(gpos4)
#> 12179072 bytes
object.size(gr4)     # 8 times bigger than the StitchedGPos object!
#> 109481200 bytes

## Shuffling the order of the positions impacts memory usage:
gpos4r <- rev(gpos4)
object.size(gpos4r)  # significantly
#> 109482784 bytes
gpos4s <- sample(gpos4)
object.size(gpos4s)  # even worse!
#> 199548064 bytes

## If one anticipates a lot of shuffling of the genomic positions,
## then an UnstitchedGPos object should be used instead:
gpos4b <- as(gpos4, "UnstitchedGPos")
object.size(gpos4b)  # initial size is bigger than stitched version
#> 60829072 bytes
object.size(rev(gpos4b))  # size didn't change
#> 60829072 bytes
object.size(sample(gpos4b))  # size increased, but is still < stitched
#> 150898784 bytes
                             # version

## AN IMPORTANT NOTE: In the worst situations, GPos still performs as
## good as a GRanges object.
object.size(as(gpos4r, "GRanges"))  # same size as 'gpos4r'
#> 109481200 bytes
object.size(as(gpos4s, "GRanges"))  # same size as 'gpos4s'
#> 199546560 bytes

## Best case scenario is when the object is strictly sorted (i.e.
## positions are in strict ascending order).
## This can be checked with:
is.unsorted(gpos4, strict=TRUE)  # 'gpos4' is strictly sorted
#> [1] FALSE

## ---------------------------------------------------------------------
## USING MEMORY-EFFICIENT METADATA COLUMNS
## ---------------------------------------------------------------------
## In order to keep memory usage as low as possible, it is recommended
## to use a memory-efficient representation of the metadata columns that
## we want to set on the object. Rle's are particularly well suited for
## this, especially if the metadata columns contain long runs of
## identical values. This is the case for example if we want to use a
## GPos object to represent the coverage of sequencing reads along a
## genome.

## Example 5:
library(pasillaBamSubset)
library(Rsamtools)  # for the BamFile() constructor function
bamfile1 <- BamFile(untreated1_chr4())
bamfile2 <- BamFile(untreated3_chr4())
gpos5 <- GPos(seqinfo(bamfile1))
library(GenomicAlignments)  # for "coverage" method for BamFile objects
#> Loading required package: SummarizedExperiment
#> Loading required package: MatrixGenerics
#> Loading required package: matrixStats
#> 
#> Attaching package: ‘matrixStats’
#> The following objects are masked from ‘package:Biobase’:
#> 
#>     anyMissing, rowMedians
#> 
#> Attaching package: ‘MatrixGenerics’
#> The following objects are masked from ‘package:matrixStats’:
#> 
#>     colAlls, colAnyNAs, colAnys, colAvgsPerRowSet, colCollapse,
#>     colCounts, colCummaxs, colCummins, colCumprods, colCumsums,
#>     colDiffs, colIQRDiffs, colIQRs, colLogSumExps, colMadDiffs,
#>     colMads, colMaxs, colMeans2, colMedians, colMins, colOrderStats,
#>     colProds, colQuantiles, colRanges, colRanks, colSdDiffs, colSds,
#>     colSums2, colTabulates, colVarDiffs, colVars, colWeightedMads,
#>     colWeightedMeans, colWeightedMedians, colWeightedSds,
#>     colWeightedVars, rowAlls, rowAnyNAs, rowAnys, rowAvgsPerColSet,
#>     rowCollapse, rowCounts, rowCummaxs, rowCummins, rowCumprods,
#>     rowCumsums, rowDiffs, rowIQRDiffs, rowIQRs, rowLogSumExps,
#>     rowMadDiffs, rowMads, rowMaxs, rowMeans2, rowMedians, rowMins,
#>     rowOrderStats, rowProds, rowQuantiles, rowRanges, rowRanks,
#>     rowSdDiffs, rowSds, rowSums2, rowTabulates, rowVarDiffs, rowVars,
#>     rowWeightedMads, rowWeightedMeans, rowWeightedMedians,
#>     rowWeightedSds, rowWeightedVars
#> The following object is masked from ‘package:Biobase’:
#> 
#>     rowMedians
cvg1 <- unlist(coverage(bamfile1), use.names=FALSE)
cvg2 <- unlist(coverage(bamfile2), use.names=FALSE)
mcols(gpos5) <- DataFrame(cvg1, cvg2)
gpos5
#> StitchedGPos object with 120748101 positions and 2 metadata columns:
#>               seqnames       pos strand |  cvg1  cvg2
#>                  <Rle> <integer>  <Rle> | <Rle> <Rle>
#>           [1]    chr2L         1      * |     0     0
#>           [2]    chr2L         2      * |     0     0
#>           [3]    chr2L         3      * |     0     0
#>           [4]    chr2L         4      * |     0     0
#>           [5]    chr2L         5      * |     0     0
#>           ...      ...       ...    ... .   ...   ...
#>   [120748097]  chrYHet    347034      * |     0     0
#>   [120748098]  chrYHet    347035      * |     0     0
#>   [120748099]  chrYHet    347036      * |     0     0
#>   [120748100]  chrYHet    347037      * |     0     0
#>   [120748101]  chrYHet    347038      * |     0     0
#>   -------
#>   seqinfo: 8 sequences from an unspecified genome

object.size(gpos5)  # lightweight
#> 1828208 bytes

## Keep only the positions where coverage is at least 10 in one of the
## 2 samples:
gpos5[mcols(gpos5)$cvg1 >= 10 | mcols(gpos5)$cvg2 >= 10]
#> StitchedGPos object with 214773 positions and 2 metadata columns:
#>            seqnames       pos strand |  cvg1  cvg2
#>               <Rle> <integer>  <Rle> | <Rle> <Rle>
#>        [1]     chr4      4936      * |    11     6
#>        [2]     chr4      4937      * |    12     6
#>        [3]     chr4      4938      * |    12     6
#>        [4]     chr4      4939      * |    13     6
#>        [5]     chr4      4940      * |    13     6
#>        ...      ...       ...    ... .   ...   ...
#>   [214769]     chr4   1348339      * |    10    14
#>   [214770]     chr4   1348340      * |     6    12
#>   [214771]     chr4   1348341      * |     6    10
#>   [214772]     chr4   1348342      * |     6    10
#>   [214773]     chr4   1348343      * |     0    10
#>   -------
#>   seqinfo: 8 sequences from an unspecified genome

## ---------------------------------------------------------------------
## USING A GPos OBJECT IN A RangedSummarizedExperiment OBJECT
## ---------------------------------------------------------------------
## Because the GPos class extends the GenomicRanges virtual class, a
## GPos derivative can be used as the rowRanges component of a
## RangedSummarizedExperiment object.

## As a 1st example, we show how the counts for samples sA and sB in
## 'gpos3' can be stored in a SummarizedExperiment object where the rows
## correspond to unique genomic positions and the columns to samples:
library(SummarizedExperiment)
counts <- cbind(sA=sA_counts, sB=sB_counts)
mcols(gpos3A) <- NULL
rse3 <- SummarizedExperiment(list(counts=counts), rowRanges=gpos3A)
rse3
#> class: RangedSummarizedExperiment 
#> dim: 1996 2 
#> metadata(0):
#> assays(1): counts
#> rownames: NULL
#> rowData names(0):
#> colnames(2): sA sB
#> colData names(0):
rowRanges(rse3)
#> StitchedGPos object with 1996 positions and 0 metadata columns:
#>          seqnames       pos strand
#>             <Rle> <integer>  <Rle>
#>      [1]     chrI         1      *
#>      [2]     chrI         2      *
#>      [3]     chrI         3      *
#>      [4]     chrI         4      *
#>      [5]     chrI         5      *
#>      ...      ...       ...    ...
#>   [1992]     chrI      1996      *
#>   [1993]     chrI      1997      *
#>   [1994]     chrI      1998      *
#>   [1995]     chrI      1999      *
#>   [1996]     chrI      2000      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
head(assay(rse3))
#>      sA sB
#> [1,]  9  9
#> [2,]  9  4
#> [3,] 10  3
#> [4,] 10  1
#> [5,]  7  7
#> [6,]  2  2

## Finally we show how the coverage data from Example 5 can be easily
## stored in a lightweight SummarizedExperiment derivative:
cvg <- mcols(gpos5)
mcols(gpos5) <- NULL
rse5 <- SummarizedExperiment(list(cvg=cvg), rowRanges=gpos5)
rse5
#> class: RangedSummarizedExperiment 
#> dim: 120748101 2 
#> metadata(0):
#> assays(1): cvg
#> rownames: NULL
#> rowData names(0):
#> colnames(2): cvg1 cvg2
#> colData names(0):
rowRanges(rse5)
#> StitchedGPos object with 120748101 positions and 0 metadata columns:
#>               seqnames       pos strand
#>                  <Rle> <integer>  <Rle>
#>           [1]    chr2L         1      *
#>           [2]    chr2L         2      *
#>           [3]    chr2L         3      *
#>           [4]    chr2L         4      *
#>           [5]    chr2L         5      *
#>           ...      ...       ...    ...
#>   [120748097]  chrYHet    347034      *
#>   [120748098]  chrYHet    347035      *
#>   [120748099]  chrYHet    347036      *
#>   [120748100]  chrYHet    347037      *
#>   [120748101]  chrYHet    347038      *
#>   -------
#>   seqinfo: 8 sequences from an unspecified genome
assay(rse5)
#> DataFrame with 120748101 rows and 2 columns
#>            cvg1  cvg2
#>           <Rle> <Rle>
#> 1             0     0
#> 2             0     0
#> 3             0     0
#> 4             0     0
#> 5             0     0
#> ...         ...   ...
#> 120748097     0     0
#> 120748098     0     0
#> 120748099     0     0
#> 120748100     0     0
#> 120748101     0     0

## Keep only the positions where coverage is at least 10 in one of the
## 2 samples:
rse5[assay(rse5)$cvg1 >= 10 | assay(rse5)$cvg2 >= 10]
#> class: RangedSummarizedExperiment 
#> dim: 214773 2 
#> metadata(0):
#> assays(1): cvg
#> rownames: NULL
#> rowData names(0):
#> colnames(2): cvg1 cvg2
#> colData names(0):