GRangesFactor objects

A GRangesFactor object is a Factor derivative where the levels are a GRanges object.

See ?Factor and in the S4Vectors package for general information about Factor objects.

GRangesFactor(x, levels, index=NULL, ...)  # constructor function

Arguments

x, levels

Like with the Factor() constructor function, at least one of x and levels must be specified. If index is NULL, both can be specified.

When x and/or levels are specified, they must be GRanges objects or derivatives. In addition, levels cannot contain duplicate ranges (i.e. anyDuplicated(levels) must return 0).

When x and levels are both specified, they should both be GRanges objects or GRanges derivatives of the same class, and all the elements in x must be represented in levels (i.e. the integer vector returned by match(x, levels) should contain no NAs).

index

NULL or an integer (or numeric) vector of valid positive indices (no NAs) into levels.

...

Optional metadata columns.

Details

Like with the Factor() constructor function, there are 4 different ways to use the GRangesFactor() constructor function. See Details section in the man page for Factor objects for more information.

Value

A GRangesFactor object.

Accessors

GRangesFactor objects support the accessors documented in the man page for Factor objects.

In addition, the following getters are supported for convenience: seqnames(), start(), end(), width(), strand(), seqinfo(), granges(), and ranges(). When called on GRangesFactor object x, they all behave as if they were called on unfactor(x).

Decoding a Factor

Because a GRangesFactor object x is a Factor derivative, unfactor(x) can be used to decode it. unfactor(x) returns an object of the same class as levels(x) (i.e. a GRanges object or derivative) and same length as x.

See ?unfactor for more information.

Coercion

GRangesFactor objects support the coercions documented in the man page for Factor objects.

In addition, coercion back and forth between GRanges and GRangesFactor is supported via as(x, "GRanges") and as(x, "GRangesFactor").

Subsetting

A GRangesFactor object can be subsetted with [, like a Factor object.

Concatenation

2 or more GRangesFactor objects can be concatenated with c(). The result of this concatenation is another GRangesFactor object.

See Concatenation section in ?Factor.

Comparing & ordering

See Comparing & Ordering section in ?Factor.

Author

Hervé Pagès

See also

Examples

showClass("GRangesFactor")  # GRangesFactor extends Factor
#> Class "GRangesFactor" [package "GenomicRanges"]
#> 
#> Slots:
#>                                                                       
#> Name:           levels elementMetadata           index        metadata
#> Class:         GRanges       DataFrame  integer_OR_raw            list
#> 
#> Extends: 
#> Class "Factor", directly
#> Class "Vector", by class "Factor", distance 2
#> Class "Annotated", by class "Factor", distance 3
#> Class "vector_OR_Vector", by class "Factor", distance 3

## ---------------------------------------------------------------------
## CONSTRUCTOR & ACCESSORS
## ---------------------------------------------------------------------
set.seed(123)
ir0 <- IRanges(sample(5, 8, replace=TRUE), width=10, names=letters[1:8])
gr0 <- GRanges("chrA", ir0, ID=paste0("ID", 1:8))

## Use explicit levels:
gr1 <- GRanges("chrA", IRanges(1:6, width=10))
grf1 <- GRangesFactor(gr0, levels=gr1)
grf1
#> GRangesFactor object of length 8 with 1 metadata column:
#>     seqnames    ranges strand |          ID
#>        <Rle> <IRanges>  <Rle> | <character>
#>   a     chrA      3-12      * |         ID1
#>   b     chrA      3-12      * |         ID2
#>   c     chrA      2-11      * |         ID3
#>   d     chrA      2-11      * |         ID4
#>   e     chrA      3-12      * |         ID5
#>   f     chrA      5-14      * |         ID6
#>   g     chrA      4-13      * |         ID7
#>   h     chrA      1-10      * |         ID8
#> 6 levels:
#>   [1] chrA:1-10 chrA:2-11 chrA:3-12 chrA:4-13 chrA:5-14 chrA:6-15
length(grf1)
#> [1] 8
names(grf1)
#> [1] "a" "b" "c" "d" "e" "f" "g" "h"
levels(grf1)  # gr1
#> GRanges object with 6 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chrA      1-10      *
#>   [2]     chrA      2-11      *
#>   [3]     chrA      3-12      *
#>   [4]     chrA      4-13      *
#>   [5]     chrA      5-14      *
#>   [6]     chrA      6-15      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
nlevels(grf1)
#> [1] 6
as.integer(grf1)  # encoding
#> a b c d e f g h 
#> 3 3 2 2 3 5 4 1 

## If we don't specify the levels, they'll be set to unique(gr0):
grf2 <- GRangesFactor(gr0)
grf2
#> GRangesFactor object of length 8 with 1 metadata column:
#>     seqnames    ranges strand |          ID
#>        <Rle> <IRanges>  <Rle> | <character>
#>   a     chrA      3-12      * |         ID1
#>   b     chrA      3-12      * |         ID2
#>   c     chrA      2-11      * |         ID3
#>   d     chrA      2-11      * |         ID4
#>   e     chrA      3-12      * |         ID5
#>   f     chrA      5-14      * |         ID6
#>   g     chrA      4-13      * |         ID7
#>   h     chrA      1-10      * |         ID8
#> 5 levels:
#>   [1] chrA:3-12 chrA:2-11 chrA:5-14 chrA:4-13 chrA:1-10
length(grf2)
#> [1] 8
names(grf2)
#> [1] "a" "b" "c" "d" "e" "f" "g" "h"
levels(grf2)  # unique(gr0)
#> GRanges object with 5 ranges and 1 metadata column:
#>     seqnames    ranges strand |          ID
#>        <Rle> <IRanges>  <Rle> | <character>
#>   a     chrA      3-12      * |         ID1
#>   c     chrA      2-11      * |         ID3
#>   f     chrA      5-14      * |         ID6
#>   g     chrA      4-13      * |         ID7
#>   h     chrA      1-10      * |         ID8
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
nlevels(grf2)
#> [1] 5
as.integer(grf2)
#> a b c d e f g h 
#> 1 1 2 2 1 3 4 5 

## ---------------------------------------------------------------------
## DECODING
## ---------------------------------------------------------------------
unfactor(grf1)
#> GRanges object with 8 ranges and 1 metadata column:
#>     seqnames    ranges strand |          ID
#>        <Rle> <IRanges>  <Rle> | <character>
#>   a     chrA      3-12      * |         ID1
#>   b     chrA      3-12      * |         ID2
#>   c     chrA      2-11      * |         ID3
#>   d     chrA      2-11      * |         ID4
#>   e     chrA      3-12      * |         ID5
#>   f     chrA      5-14      * |         ID6
#>   g     chrA      4-13      * |         ID7
#>   h     chrA      1-10      * |         ID8
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

stopifnot(identical(gr0, unfactor(grf1)))
stopifnot(identical(gr0, unfactor(grf2)))

unfactor(grf1, use.names=FALSE)
#> GRanges object with 8 ranges and 1 metadata column:
#>       seqnames    ranges strand |          ID
#>          <Rle> <IRanges>  <Rle> | <character>
#>   [1]     chrA      3-12      * |         ID1
#>   [2]     chrA      3-12      * |         ID2
#>   [3]     chrA      2-11      * |         ID3
#>   [4]     chrA      2-11      * |         ID4
#>   [5]     chrA      3-12      * |         ID5
#>   [6]     chrA      5-14      * |         ID6
#>   [7]     chrA      4-13      * |         ID7
#>   [8]     chrA      1-10      * |         ID8
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
unfactor(grf1, ignore.mcols=TRUE)
#> GRanges object with 8 ranges and 0 metadata columns:
#>     seqnames    ranges strand
#>        <Rle> <IRanges>  <Rle>
#>   a     chrA      3-12      *
#>   b     chrA      3-12      *
#>   c     chrA      2-11      *
#>   d     chrA      2-11      *
#>   e     chrA      3-12      *
#>   f     chrA      5-14      *
#>   g     chrA      4-13      *
#>   h     chrA      1-10      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## ---------------------------------------------------------------------
## COERCION
## ---------------------------------------------------------------------
grf2b <- as(gr0, "GRangesFactor")  # same as GRangesFactor(gr0)
stopifnot(identical(grf2, grf2b))

as.factor(grf2)
#>         a         b         c         d         e         f         g         h 
#> chrA:3-12 chrA:3-12 chrA:2-11 chrA:2-11 chrA:3-12 chrA:5-14 chrA:4-13 chrA:1-10 
#> Levels: chrA:3-12 chrA:2-11 chrA:5-14 chrA:4-13 chrA:1-10
as.factor(grf1)
#>         a         b         c         d         e         f         g         h 
#> chrA:3-12 chrA:3-12 chrA:2-11 chrA:2-11 chrA:3-12 chrA:5-14 chrA:4-13 chrA:1-10 
#> Levels: chrA:1-10 chrA:2-11 chrA:3-12 chrA:4-13 chrA:5-14 chrA:6-15

as.character(grf1)  # same as unfactor(as.factor(grf1)),
#>           a           b           c           d           e           f 
#> "chrA:3-12" "chrA:3-12" "chrA:2-11" "chrA:2-11" "chrA:3-12" "chrA:5-14" 
#>           g           h 
#> "chrA:4-13" "chrA:1-10" 
                    # and also same as as.character(unfactor(grf1))

## ---------------------------------------------------------------------
## CONCATENATION
## ---------------------------------------------------------------------
gr3 <- GRanges("chrA", IRanges(c(5, 2, 8:6), width=10))
grf3 <- GRangesFactor(levels=gr3, index=2:4)
grf13 <- c(grf1, grf3)
grf13
#> GRangesFactor object of length 11 with 1 metadata column:
#>     seqnames    ranges strand |          ID
#>        <Rle> <IRanges>  <Rle> | <character>
#>   a     chrA      3-12      * |         ID1
#>   b     chrA      3-12      * |         ID2
#>   c     chrA      2-11      * |         ID3
#>   d     chrA      2-11      * |         ID4
#>   e     chrA      3-12      * |         ID5
#>   f     chrA      5-14      * |         ID6
#>   g     chrA      4-13      * |         ID7
#>   h     chrA      1-10      * |         ID8
#>         chrA      2-11      * |        <NA>
#>         chrA      8-17      * |        <NA>
#>         chrA      7-16      * |        <NA>
#> 8 levels:
#>   [1] chrA:1-10 chrA:2-11 chrA:3-12 chrA:4-13 chrA:5-14 chrA:6-15 chrA:8-17
#>   [8] chrA:7-16
levels(grf13)
#> GRanges object with 8 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chrA      1-10      *
#>   [2]     chrA      2-11      *
#>   [3]     chrA      3-12      *
#>   [4]     chrA      4-13      *
#>   [5]     chrA      5-14      *
#>   [6]     chrA      6-15      *
#>   [7]     chrA      8-17      *
#>   [8]     chrA      7-16      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

stopifnot(identical(c(unfactor(grf1), unfactor(grf3)), unfactor(grf13)))

## ---------------------------------------------------------------------
## COMPARING & ORDERING
## ---------------------------------------------------------------------
grf1 == grf2  # same as unfactor(grf1) == unfactor(grf2)
#> [1] TRUE TRUE TRUE TRUE TRUE TRUE TRUE TRUE

order(grf1)   # same as order(unfactor(grf1))
#> [1] 8 3 4 1 2 5 7 6
order(grf2)   # same as order(unfactor(grf2))
#> [1] 8 3 4 1 2 5 7 6

## The levels of the GRangesFactor influence the order of the table:
table(grf1)
#> grf1
#> chrA:1-10 chrA:2-11 chrA:3-12 chrA:4-13 chrA:5-14 
#>         1         2         3         1         1 
table(grf2)
#> grf2
#> chrA:1-10 chrA:2-11 chrA:3-12 chrA:4-13 chrA:5-14 
#>         1         2         3         1         1