Finding overlapping genomic ranges

Various methods for finding/counting overlaps between objects containing genomic ranges. This man page describes the methods that operate on GenomicRanges and GRangesList objects.

NOTE: The findOverlaps generic function and methods for IntegerRanges and IntegerRangesList objects are defined and documented in the IRanges package. The methods for GAlignments, GAlignmentPairs, and GAlignmentsList objects are defined and documented in the GenomicAlignments package.

GenomicRanges and GRangesList objects also support countOverlaps, overlapsAny, and subsetByOverlaps thanks to the default methods defined in the IRanges package and to the findOverlaps and countOverlaps methods defined in this package and documented below.

# S4 method for GenomicRanges,GenomicRanges
findOverlaps(query, subject,
    maxgap=-1L, minoverlap=0L,
    type=c("any", "start", "end", "within", "equal"),
    select=c("all", "first", "last", "arbitrary"),
    ignore.strand=FALSE)

# S4 method for GRangesList,GenomicRanges
findOverlaps(query, subject,
    maxgap=-1L, minoverlap=0L,
    type=c("any", "start", "end", "within", "equal"),
    select=c("all", "first", "last", "arbitrary"),
    ignore.strand=FALSE)

# S4 method for GenomicRanges,GRangesList
findOverlaps(query, subject,
    maxgap=-1L, minoverlap=0L,
    type=c("any", "start", "end", "within", "equal"),
    select=c("all", "first", "last", "arbitrary"),
    ignore.strand=FALSE)

# S4 method for GRangesList,GRangesList
findOverlaps(query, subject,
    maxgap=-1L, minoverlap=0L,
    type=c("any", "start", "end", "within", "equal"),
    select=c("all", "first", "last", "arbitrary"),
    ignore.strand=FALSE)

# S4 method for GenomicRanges,GenomicRanges
countOverlaps(query, subject,
    maxgap=-1L, minoverlap=0L,
    type=c("any", "start", "end", "within", "equal"),
    ignore.strand=FALSE)

Arguments

query, subject

A GRanges or GRangesList object.

maxgap, minoverlap, type

See ?findOverlaps in the IRanges package for a description of these arguments.

IMPORTANT NOTE about how minoverlap is interpreted when query or subject is a GRangesList object: In this case, the total number of overlapping positions between a given element in query and a given element in subject is taken into account. For example, if query is a GRanges object, and subject a GRangesList object, then findOverlaps() will report an overlap between query[i] (a single range) and subject[[j]] (multiple ranges) only if the total number of positions in subject[[j]] that overlap with query[i] is equal to minoverlap or greater. In other words, the full overlap across all the ranges in subject[[j]] is looked at. See the Examples section below for an example illustrating this.

select

When select is "all" (the default), the results are returned as a Hits object. Otherwise the returned value is an integer vector parallel to query (i.e. same length) containing the first, last, or arbitrary overlapping interval in subject, with NA indicating intervals that did not overlap any intervals in subject.

ignore.strand

When set to TRUE, the strand information is ignored in the overlap calculations.

Details

When the query and the subject are GRanges or GRangesList objects, findOverlaps uses the triplet (sequence name, range, strand) to determine which features (see paragraph below for the definition of feature) from the query overlap which features in the subject, where a strand value of "*" is treated as occurring on both the "+" and "-" strand. An overlap is recorded when a feature in the query and a feature in the subject have the same sequence name, have a compatible pairing of strands (e.g. "+"/"+", "-"/"-", "*"/"+", "*"/"-", etc.), and satisfy the interval overlap requirements.

In the context of findOverlaps, a feature is a collection of ranges that are treated as a single entity. For GRanges objects, a feature is a single range; while for GRangesList objects, a feature is a list element containing a set of ranges. In the results, the features are referred to by number, which run from 1 to length(query)/length(subject).

For type="equal" with GRangesList objects, query[[i]] matches subject[[j]] iff for each range in query[[i]] there is an identical range in subject[[j]], and vice versa.

Value

For findOverlaps: either a Hits object when

select="all" or an integer vector otherwise.

For countOverlaps: an integer vector containing the tabulated query overlap hits.

Author

P. Aboyoun, S. Falcon, M. Lawrence, and H. Pagès

See also

Examples

## ---------------------------------------------------------------------
## BASIC EXAMPLES
## ---------------------------------------------------------------------

## GRanges object:
gr <- GRanges(
        seqnames=Rle(c("chr1", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
        ranges=IRanges(1:10, width=10:1, names=head(letters,10)),
        strand=Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
        score=1:10,
        GC=seq(1, 0, length=10)
      )
gr
#> GRanges object with 10 ranges and 2 metadata columns:
#>     seqnames    ranges strand |     score        GC
#>        <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   a     chr1      1-10      - |         1  1.000000
#>   b     chr2      2-10      + |         2  0.888889
#>   c     chr2      3-10      + |         3  0.777778
#>   d     chr2      4-10      * |         4  0.666667
#>   e     chr1      5-10      * |         5  0.555556
#>   f     chr1      6-10      + |         6  0.444444
#>   g     chr3      7-10      + |         7  0.333333
#>   h     chr3      8-10      + |         8  0.222222
#>   i     chr3      9-10      - |         9  0.111111
#>   j     chr3        10      - |        10  0.000000
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

## GRangesList object:
gr1 <- GRanges(seqnames="chr2", ranges=IRanges(4:3, 6),
               strand="+", score=5:4, GC=0.45)
gr2 <- GRanges(seqnames=c("chr1", "chr1"),
               ranges=IRanges(c(7,13), width=3),
               strand=c("+", "-"), score=3:4, GC=c(0.3, 0.5))
gr3 <- GRanges(seqnames=c("chr1", "chr2"),
               ranges=IRanges(c(1, 4), c(3, 9)),
               strand=c("-", "-"), score=c(6L, 2L), GC=c(0.4, 0.1))
grl <- GRangesList("gr1"=gr1, "gr2"=gr2, "gr3"=gr3)

## Overlapping two GRanges objects:
table(!is.na(findOverlaps(gr, gr1, select="arbitrary")))
#> 
#> FALSE  TRUE 
#>     7     3 
countOverlaps(gr, gr1)
#> a b c d e f g h i j 
#> 0 2 2 2 0 0 0 0 0 0 
findOverlaps(gr, gr1)
#> Hits object with 6 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         2           2
#>   [2]         2           1
#>   [3]         3           2
#>   [4]         3           1
#>   [5]         4           2
#>   [6]         4           1
#>   -------
#>   queryLength: 10 / subjectLength: 2
subsetByOverlaps(gr, gr1)
#> GRanges object with 3 ranges and 2 metadata columns:
#>     seqnames    ranges strand |     score        GC
#>        <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   b     chr2      2-10      + |         2  0.888889
#>   c     chr2      3-10      + |         3  0.777778
#>   d     chr2      4-10      * |         4  0.666667
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

countOverlaps(gr, gr1, type="start")
#> a b c d e f g h i j 
#> 0 0 1 1 0 0 0 0 0 0 
findOverlaps(gr, gr1, type="start")
#> Hits object with 2 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         3           2
#>   [2]         4           1
#>   -------
#>   queryLength: 10 / subjectLength: 2
subsetByOverlaps(gr, gr1, type="start")
#> GRanges object with 2 ranges and 2 metadata columns:
#>     seqnames    ranges strand |     score        GC
#>        <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   c     chr2      3-10      + |         3  0.777778
#>   d     chr2      4-10      * |         4  0.666667
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

findOverlaps(gr, gr1, select="first")
#>  [1] NA  1  1  1 NA NA NA NA NA NA
findOverlaps(gr, gr1, select="last")
#>  [1] NA  2  2  2 NA NA NA NA NA NA

findOverlaps(gr1, gr)
#> Hits object with 6 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           2
#>   [2]         1           3
#>   [3]         1           4
#>   [4]         2           2
#>   [5]         2           3
#>   [6]         2           4
#>   -------
#>   queryLength: 2 / subjectLength: 10
findOverlaps(gr1, gr, type="start")
#> Hits object with 2 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           4
#>   [2]         2           3
#>   -------
#>   queryLength: 2 / subjectLength: 10
findOverlaps(gr1, gr, type="within")
#> Hits object with 5 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           2
#>   [2]         1           3
#>   [3]         1           4
#>   [4]         2           2
#>   [5]         2           3
#>   -------
#>   queryLength: 2 / subjectLength: 10
findOverlaps(gr1, gr, type="equal")
#> Hits object with 0 hits and 0 metadata columns:
#>    queryHits subjectHits
#>    <integer>   <integer>
#>   -------
#>   queryLength: 2 / subjectLength: 10

## ---------------------------------------------------------------------
## MORE EXAMPLES
## ---------------------------------------------------------------------

table(!is.na(findOverlaps(gr, gr1, select="arbitrary")))
#> 
#> FALSE  TRUE 
#>     7     3 
countOverlaps(gr, gr1)
#> a b c d e f g h i j 
#> 0 2 2 2 0 0 0 0 0 0 
findOverlaps(gr, gr1)
#> Hits object with 6 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         2           2
#>   [2]         2           1
#>   [3]         3           2
#>   [4]         3           1
#>   [5]         4           2
#>   [6]         4           1
#>   -------
#>   queryLength: 10 / subjectLength: 2
subsetByOverlaps(gr, gr1)
#> GRanges object with 3 ranges and 2 metadata columns:
#>     seqnames    ranges strand |     score        GC
#>        <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   b     chr2      2-10      + |         2  0.888889
#>   c     chr2      3-10      + |         3  0.777778
#>   d     chr2      4-10      * |         4  0.666667
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

## Overlaps between a GRanges and a GRangesList object:

table(!is.na(findOverlaps(grl, gr, select="first")))
#> 
#> TRUE 
#>    3 
countOverlaps(grl, gr)
#> gr1 gr2 gr3 
#>   3   2   2 
findOverlaps(grl, gr)
#> Hits object with 7 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           2
#>   [2]         1           3
#>   [3]         1           4
#>   [4]         2           5
#>   [5]         2           6
#>   [6]         3           1
#>   [7]         3           4
#>   -------
#>   queryLength: 3 / subjectLength: 10
subsetByOverlaps(grl, gr)
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr2       4-6      + |         5      0.45
#>   [2]     chr2       3-6      + |         4      0.45
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr1       7-9      + |         3       0.3
#>   [2]     chr1     13-15      - |         4       0.5
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr1       1-3      - |         6       0.4
#>   [2]     chr2       4-9      - |         2       0.1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
countOverlaps(grl, gr, type="start")
#> gr1 gr2 gr3 
#>   2   0   2 
findOverlaps(grl, gr, type="start")
#> Hits object with 4 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           4
#>   [2]         1           3
#>   [3]         3           1
#>   [4]         3           4
#>   -------
#>   queryLength: 3 / subjectLength: 10
subsetByOverlaps(grl, gr, type="start")
#> GRangesList object of length 2:
#> $gr1
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr2       4-6      + |         5      0.45
#>   [2]     chr2       3-6      + |         4      0.45
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr1       1-3      - |         6       0.4
#>   [2]     chr2       4-9      - |         2       0.1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
findOverlaps(grl, gr, select="first")
#> [1] 2 5 1

table(!is.na(findOverlaps(grl, gr1, select="first")))
#> 
#> FALSE  TRUE 
#>     2     1 
countOverlaps(grl, gr1)
#> gr1 gr2 gr3 
#>   2   0   0 
findOverlaps(grl, gr1)
#> Hits object with 2 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           2
#>   [2]         1           1
#>   -------
#>   queryLength: 3 / subjectLength: 2
subsetByOverlaps(grl, gr1)
#> GRangesList object of length 1:
#> $gr1
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr2       4-6      + |         5      0.45
#>   [2]     chr2       3-6      + |         4      0.45
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
countOverlaps(grl, gr1, type="start")
#> gr1 gr2 gr3 
#>   2   0   0 
findOverlaps(grl, gr1, type="start")
#> Hits object with 2 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           1
#>   [2]         1           2
#>   -------
#>   queryLength: 3 / subjectLength: 2
subsetByOverlaps(grl, gr1, type="start")
#> GRangesList object of length 1:
#> $gr1
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr2       4-6      + |         5      0.45
#>   [2]     chr2       3-6      + |         4      0.45
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
findOverlaps(grl, gr1, select="first")
#> [1]  1 NA NA

## Overlaps between two GRangesList objects:
countOverlaps(grl, rev(grl))
#> gr1 gr2 gr3 
#>   1   1   1 
findOverlaps(grl, rev(grl))
#> Hits object with 3 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           3
#>   [2]         2           2
#>   [3]         3           1
#>   -------
#>   queryLength: 3 / subjectLength: 3
subsetByOverlaps(grl, rev(grl))
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr2       4-6      + |         5      0.45
#>   [2]     chr2       3-6      + |         4      0.45
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr1       7-9      + |         3       0.3
#>   [2]     chr1     13-15      - |         4       0.5
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr1       1-3      - |         6       0.4
#>   [2]     chr2       4-9      - |         2       0.1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 

## ---------------------------------------------------------------------
## INTERPRETATION OF 'minoverlap' WHEN 'query' OR 'subject' IS A
## GRangesList OBJECT
## ---------------------------------------------------------------------
gr1 <- GRanges("chr5:1-26")
gr2 <- GRanges("chr5:31-40")
gr3 <- c(GRanges("chr5:11-20"), gr2)
grl123 <- GRangesList(gr1, gr2, gr3)
grl123
#> GRangesList object of length 3:
#> [[1]]
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr5      1-26      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> [[2]]
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr5     31-40      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 
#> [[3]]
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr5     11-20      *
#>   [2]     chr5     31-40      *
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths
#> 

query <- GRanges("chr5:17-35")

findOverlaps(query, grl123[[1]], minoverlap=8)  # 1 hit
#> Hits object with 1 hit and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           1
#>   -------
#>   queryLength: 1 / subjectLength: 1
findOverlaps(query, grl123[[2]], minoverlap=8)  # no hit
#> Hits object with 0 hits and 0 metadata columns:
#>    queryHits subjectHits
#>    <integer>   <integer>
#>   -------
#>   queryLength: 1 / subjectLength: 1
findOverlaps(query, grl123[[3]], minoverlap=8)  # no hit
#> Hits object with 0 hits and 0 metadata columns:
#>    queryHits subjectHits
#>    <integer>   <integer>
#>   -------
#>   queryLength: 1 / subjectLength: 2

## Using GRangesList object 'grl123' as the subject:
findOverlaps(query, grl123, minoverlap=8)
#> Hits object with 2 hits and 0 metadata columns:
#>       queryHits subjectHits
#>       <integer>   <integer>
#>   [1]         1           1
#>   [2]         1           3
#>   -------
#>   queryLength: 1 / subjectLength: 3

## As we can see, a hit is reported with the 3rd element in the subject.
## That's because the total number of positions in this overlap is 9:
## - positions 17 to 20 in the first range of grl123[[3]], so 4 positions
## - positions 31 to 35 in its second range, so 5 positions

## Sanity checks:
hits <- findOverlaps(query, grl123[[1]], minoverlap=8)
stopifnot(length(hits) == 1)
hits <- findOverlaps(query, grl123[[2]], minoverlap=8)
stopifnot(length(hits) == 0)
hits <- findOverlaps(query, grl123[[3]], minoverlap=8)
stopifnot(length(hits) == 0)
hits <- findOverlaps(query, grl123, minoverlap=8)
stopifnot(identical(subjectHits(hits), c(1L, 3L)))
hits <- findOverlaps(query, grl123, minoverlap=9)
stopifnot(identical(subjectHits(hits), c(1L, 3L)))
hits <- findOverlaps(query, grl123, minoverlap=10)
stopifnot(identical(subjectHits(hits), 1L))
hits <- findOverlaps(query, grl123, minoverlap=11)
stopifnot(length(hits) == 0)