Inter range transformations of a GRanges or GRangesList object

This man page documents inter range transformations of a GenomicRanges object (i.e. of an object that belongs to the GenomicRanges class or one of its subclasses, this includes for example GRanges objects), or a GRangesList object.

See ?`intra-range-methods` and ?`inter-range-methods` in the IRanges package for a quick introduction to intra range and inter range transformations.

See ?`intra-range-methods` for intra range transformations of a GenomicRanges object or GRangesList object.

# S4 method for GenomicRanges
range(x, ..., with.revmap=FALSE, ignore.strand=FALSE, na.rm=FALSE)
# S4 method for GRangesList
range(x, ..., with.revmap=FALSE, ignore.strand=FALSE, na.rm=FALSE)

# S4 method for GenomicRanges
reduce(x, drop.empty.ranges=FALSE, min.gapwidth=1L, with.revmap=FALSE,
          with.inframe.attrib=FALSE, ignore.strand=FALSE)
# S4 method for GRangesList
reduce(x, drop.empty.ranges=FALSE, min.gapwidth=1L, with.revmap=FALSE,
          with.inframe.attrib=FALSE, ignore.strand=FALSE)

# S4 method for GenomicRanges
gaps(x, start=1L, end=seqlengths(x))

# S4 method for GenomicRanges
disjoin(x, with.revmap=FALSE, ignore.strand=FALSE)
# S4 method for GRangesList
disjoin(x, with.revmap=FALSE, ignore.strand=FALSE)

# S4 method for GenomicRanges
isDisjoint(x, ignore.strand=FALSE)
# S4 method for GRangesList
isDisjoint(x, ignore.strand=FALSE)

# S4 method for GenomicRanges
disjointBins(x, ignore.strand=FALSE)

Arguments

x

A GenomicRanges or GenomicRangesList object.

drop.empty.ranges, min.gapwidth, with.revmap, with.inframe.attrib, start, end

See ?`inter-range-methods` in the IRanges package.

ignore.strand

TRUE or FALSE. Whether the strand of the input ranges should be ignored or not. See details below.

...

For range, additional GenomicRanges objects to consider. Ignored otherwise.

na.rm

Ignored.

Details

On a GRanges object

range returns an object of the same type as x containing range bounds for each distinct (seqname, strand) pairing. The names (names(x)) and the metadata columns in x are dropped.

reduce returns an object of the same type as x containing reduced ranges for each distinct (seqname, strand) pairing. The names (names(x)) and the metadata columns in x are dropped. See ?reduce for more information about range reduction and for a description of the optional arguments.

gaps returns an object of the same type as x containing complemented ranges for each distinct (seqname, strand) pairing. The names (names(x)) and the columns in x are dropped. For the start and end arguments of this gaps method, it is expected that the user will supply a named integer vector (where the names correspond to the appropriate seqlevels). See ?gaps for more information about range complements and for a description of the optional arguments.

disjoin returns an object of the same type as x containing disjoint ranges for each distinct (seqname, strand) pairing. The names (names(x)) and the metadata columns in x are dropped. If with.revmap=TRUE, a metadata column that maps the ouput ranges to the input ranges is added to the returned object. See ?disjoin for more information.

isDisjoint returns a logical value indicating whether the ranges in x are disjoint (i.e. non-overlapping).

disjointBins returns bin indexes for the ranges in x, such that ranges in the same bin do not overlap. If ignore.strand=FALSE, the two features cannot overlap if they are on different strands.

On a GRangesList/GenomicRangesList object

When they are supported on GRangesList object x, the above inter range transformations will apply the transformation to each of the list elements in x and return a list-like object parallel to x (i.e. with 1 list element per list element in x). If x has names on it, they're propagated to the returned object.

Author

H. Pagès and P. Aboyoun

See also

Examples

gr <- GRanges(
        seqnames=Rle(paste("chr", c(1, 2, 1, 3), sep=""), c(1, 3, 2, 4)),
        ranges=IRanges(1:10, width=10:1, names=letters[1:10]),
        strand=Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
        score=1:10,
        GC=seq(1, 0, length=10)
      )
gr
#> GRanges object with 10 ranges and 2 metadata columns:
#>     seqnames    ranges strand |     score        GC
#>        <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   a     chr1      1-10      - |         1  1.000000
#>   b     chr2      2-10      + |         2  0.888889
#>   c     chr2      3-10      + |         3  0.777778
#>   d     chr2      4-10      * |         4  0.666667
#>   e     chr1      5-10      * |         5  0.555556
#>   f     chr1      6-10      + |         6  0.444444
#>   g     chr3      7-10      + |         7  0.333333
#>   h     chr3      8-10      + |         8  0.222222
#>   i     chr3      9-10      - |         9  0.111111
#>   j     chr3        10      - |        10  0.000000
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

gr1 <- GRanges(seqnames="chr2", ranges=IRanges(3, 6),
               strand="+", score=5L, GC=0.45)
gr2 <- GRanges(seqnames="chr1",
               ranges=IRanges(c(10, 7, 19), width=5),
               strand=c("+", "-", "+"), score=3:5, GC=c(0.3, 0.5, 0.66))
gr3 <- GRanges(seqnames=c("chr1", "chr2"),
               ranges=IRanges(c(1, 4), c(3, 9)),
               strand=c("-", "-"), score=c(6L, 2L), GC=c(0.4, 0.1))
grl <- GRangesList(gr1=gr1, gr2=gr2, gr3=gr3)
grl
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 1 range and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr2       3-6      + |         5      0.45
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 3 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr1     10-14      + |         3      0.30
#>   [2]     chr1      7-11      - |         4      0.50
#>   [3]     chr1     19-23      + |         5      0.66
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 2 metadata columns:
#>       seqnames    ranges strand |     score        GC
#>          <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   [1]     chr1       1-3      - |         6       0.4
#>   [2]     chr2       4-9      - |         2       0.1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 

## ---------------------------------------------------------------------
## range()
## ---------------------------------------------------------------------

## On a GRanges object:
range(gr)
#> GRanges object with 7 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1      6-10      +
#>   [2]     chr1      1-10      -
#>   [3]     chr1      5-10      *
#>   [4]     chr2      2-10      +
#>   [5]     chr2      4-10      *
#>   [6]     chr3      7-10      +
#>   [7]     chr3      9-10      -
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
range(gr, with.revmap=TRUE)
#> GRanges object with 7 ranges and 1 metadata column:
#>       seqnames    ranges strand |        revmap
#>          <Rle> <IRanges>  <Rle> | <IntegerList>
#>   [1]     chr1      6-10      + |             6
#>   [2]     chr1      1-10      - |             1
#>   [3]     chr1      5-10      * |             5
#>   [4]     chr2      2-10      + |           2,3
#>   [5]     chr2      4-10      * |             4
#>   [6]     chr3      7-10      + |           7,8
#>   [7]     chr3      9-10      - |          9,10
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

## On a GRangesList object:
range(grl)
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       3-6      +
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1     10-23      +
#>   [2]     chr1      7-11      -
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       4-9      -
#>   [2]     chr1       1-3      -
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
range(grl, ignore.strand=TRUE)
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       3-6      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1      7-23      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       4-9      *
#>   [2]     chr1       1-3      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
range(grl, with.revmap=TRUE, ignore.strand=TRUE)
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 1 range and 1 metadata column:
#>       seqnames    ranges strand |        revmap
#>          <Rle> <IRanges>  <Rle> | <IntegerList>
#>   [1]     chr2       3-6      * |             1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 1 range and 1 metadata column:
#>       seqnames    ranges strand |        revmap
#>          <Rle> <IRanges>  <Rle> | <IntegerList>
#>   [1]     chr1      7-23      * |         1,2,3
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 1 metadata column:
#>       seqnames    ranges strand |        revmap
#>          <Rle> <IRanges>  <Rle> | <IntegerList>
#>   [1]     chr2       4-9      * |             2
#>   [2]     chr1       1-3      * |             1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 

# ---------------------------------------------------------------------
## reduce()
## ---------------------------------------------------------------------
reduce(gr)
#> GRanges object with 7 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1      6-10      +
#>   [2]     chr1      1-10      -
#>   [3]     chr1      5-10      *
#>   [4]     chr2      2-10      +
#>   [5]     chr2      4-10      *
#>   [6]     chr3      7-10      +
#>   [7]     chr3      9-10      -
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

gr2 <- reduce(gr, with.revmap=TRUE)
revmap <- mcols(gr2)$revmap  # an IntegerList

## Use the mapping from reduced to original ranges to group the original
## ranges by reduced range:
relist(gr[unlist(revmap)], revmap)
#> GRangesList object of length 7:
#> [[1]]
#> GRanges object with 1 range and 2 metadata columns:
#>     seqnames    ranges strand |     score        GC
#>        <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   f     chr1      6-10      + |         6  0.444444
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
#> 
#> [[2]]
#> GRanges object with 1 range and 2 metadata columns:
#>     seqnames    ranges strand |     score        GC
#>        <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   a     chr1      1-10      - |         1         1
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
#> 
#> [[3]]
#> GRanges object with 1 range and 2 metadata columns:
#>     seqnames    ranges strand |     score        GC
#>        <Rle> <IRanges>  <Rle> | <integer> <numeric>
#>   e     chr1      5-10      * |         5  0.555556
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
#> 
#> ...
#> <4 more elements>

## Or use it to split the DataFrame of original metadata columns by
## reduced range:
relist(mcols(gr)[unlist(revmap), ], revmap)  # a SplitDataFrameList
#> SplitDataFrameList of length 7
#> [[1]]
#> DataFrame with 1 row and 2 columns
#>       score        GC
#>   <integer> <numeric>
#> f         6  0.444444
#> 
#> [[2]]
#> DataFrame with 1 row and 2 columns
#>       score        GC
#>   <integer> <numeric>
#> a         1         1
#> 
#> [[3]]
#> DataFrame with 1 row and 2 columns
#>       score        GC
#>   <integer> <numeric>
#> e         5  0.555556
#> 
#> [[4]]
#> DataFrame with 2 rows and 2 columns
#>       score        GC
#>   <integer> <numeric>
#> b         2  0.888889
#> c         3  0.777778
#> 
#> [[5]]
#> DataFrame with 1 row and 2 columns
#>       score        GC
#>   <integer> <numeric>
#> d         4  0.666667
#> 
#> [[6]]
#> DataFrame with 2 rows and 2 columns
#>       score        GC
#>   <integer> <numeric>
#> g         7  0.333333
#> h         8  0.222222
#> 
#> [[7]]
#> DataFrame with 2 rows and 2 columns
#>       score        GC
#>   <integer> <numeric>
#> i         9  0.111111
#> j        10  0.000000
#> 

## [For advanced users] Use this reverse mapping to compare the reduced
## ranges with the ranges they originate from:
expanded_gr2 <- rep(gr2, elementNROWS(revmap))
reordered_gr <- gr[unlist(revmap)]
codes <- pcompare(expanded_gr2, reordered_gr)
## All the codes should translate to "d", "e", "g", or "h" (the 4 letters
## indicating that the range on the left contains the range on the right):
alphacodes <- rangeComparisonCodeToLetter(pcompare(expanded_gr2, reordered_gr))
stopifnot(all(alphacodes %in% c("d", "e", "g", "h")))

## On a big GRanges object with a lot of seqlevels:
mcols(gr) <- NULL
biggr <- c(gr, GRanges("chr1", IRanges(c(4, 1), c(5, 2)), strand="+"))
seqlevels(biggr) <- paste0("chr", 1:2000)
biggr <- rep(biggr, 25000)
set.seed(33)
seqnames(biggr) <- sample(factor(seqlevels(biggr), levels=seqlevels(biggr)),
                          length(biggr), replace=TRUE)

biggr2 <- reduce(biggr, with.revmap=TRUE)
revmap <- mcols(biggr2)$revmap
expanded_biggr2 <- rep(biggr2, elementNROWS(revmap))
reordered_biggr <- biggr[unlist(revmap)]
codes <- pcompare(expanded_biggr2, reordered_biggr)
alphacodes <- rangeComparisonCodeToLetter(pcompare(expanded_biggr2,
                                                  reordered_biggr))
stopifnot(all(alphacodes %in% c("d", "e", "g", "h")))
table(alphacodes)
#> alphacodes
#>      a      b      c      d      e      f      g      h      i      j      k 
#>      0      0      0 200000  25000      0  50000  25000      0      0      0 
#>      l      m      X 
#>      0      0      0 

## On a GRangesList object:
reduce(grl)  # Doesn't really reduce anything but note the reordering
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       3-6      +
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 3 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1     10-14      +
#>   [2]     chr1     19-23      +
#>   [3]     chr1      7-11      -
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       4-9      -
#>   [2]     chr1       1-3      -
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
             # of the inner elements in the 2nd and 3rd list elements:
             # the ranges are reordered by sequence name first (which
             # should appear in the same order as in 'seqlevels(grl)'),
             # and then by strand.
reduce(grl, ignore.strand=TRUE)  # 2nd list element got reduced
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       3-6      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1      7-14      *
#>   [2]     chr1     19-23      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       4-9      *
#>   [2]     chr1       1-3      *
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 

## ---------------------------------------------------------------------
## gaps()
## ---------------------------------------------------------------------
gaps(gr, start=1, end=10)
#> GRanges object with 8 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1       1-5      +
#>   [2]     chr1       1-4      *
#>   [3]     chr2         1      +
#>   [4]     chr2      1-10      -
#>   [5]     chr2       1-3      *
#>   [6]     chr3       1-6      +
#>   [7]     chr3       1-8      -
#>   [8]     chr3      1-10      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths

## ---------------------------------------------------------------------
## disjoin(), isDisjoint(), disjointBins()
## ---------------------------------------------------------------------
disjoin(gr)
#> GRanges object with 10 ranges and 0 metadata columns:
#>        seqnames    ranges strand
#>           <Rle> <IRanges>  <Rle>
#>    [1]     chr1      6-10      +
#>    [2]     chr1      1-10      -
#>    [3]     chr1      5-10      *
#>    [4]     chr2         2      +
#>    [5]     chr2      3-10      +
#>    [6]     chr2      4-10      *
#>    [7]     chr3         7      +
#>    [8]     chr3      8-10      +
#>    [9]     chr3         9      -
#>   [10]     chr3        10      -
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
disjoin(gr, with.revmap=TRUE)
#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |        revmap
#>           <Rle> <IRanges>  <Rle> | <IntegerList>
#>    [1]     chr1      6-10      + |             6
#>    [2]     chr1      1-10      - |             1
#>    [3]     chr1      5-10      * |             5
#>    [4]     chr2         2      + |             2
#>    [5]     chr2      3-10      + |           2,3
#>    [6]     chr2      4-10      * |             4
#>    [7]     chr3         7      + |             7
#>    [8]     chr3      8-10      + |           7,8
#>    [9]     chr3         9      - |             9
#>   [10]     chr3        10      - |          9,10
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
disjoin(gr, with.revmap=TRUE, ignore.strand=TRUE)
#> GRanges object with 10 ranges and 1 metadata column:
#>        seqnames    ranges strand |        revmap
#>           <Rle> <IRanges>  <Rle> | <IntegerList>
#>    [1]     chr1       1-4      * |             1
#>    [2]     chr1         5      * |           1,5
#>    [3]     chr1      6-10      * |         1,5,6
#>    [4]     chr2         2      * |             2
#>    [5]     chr2         3      * |           2,3
#>    [6]     chr2      4-10      * |         2,3,4
#>    [7]     chr3         7      * |             7
#>    [8]     chr3         8      * |           7,8
#>    [9]     chr3         9      * |         7,8,9
#>   [10]     chr3        10      * |     7,8,9,...
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
isDisjoint(gr)
#> [1] FALSE
stopifnot(isDisjoint(disjoin(gr)))
disjointBins(gr)
#>  [1] 1 1 2 1 1 1 1 2 1 2
stopifnot(all(sapply(split(gr, disjointBins(gr)), isDisjoint)))

## On a GRangesList object:
disjoin(grl)  # doesn't really disjoin anything but note the reordering
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       3-6      +
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 3 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1     10-14      +
#>   [2]     chr1     19-23      +
#>   [3]     chr1      7-11      -
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2       4-9      -
#>   [2]     chr1       1-3      -
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
disjoin(grl, with.revmap=TRUE)
#> GRangesList object of length 3:
#> $gr1
#> GRanges object with 1 range and 1 metadata column:
#>       seqnames    ranges strand |        revmap
#>          <Rle> <IRanges>  <Rle> | <IntegerList>
#>   [1]     chr2       3-6      + |             1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr2
#> GRanges object with 3 ranges and 1 metadata column:
#>       seqnames    ranges strand |        revmap
#>          <Rle> <IRanges>  <Rle> | <IntegerList>
#>   [1]     chr1     10-14      + |             1
#>   [2]     chr1     19-23      + |             3
#>   [3]     chr1      7-11      - |             2
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#> 
#> $gr3
#> GRanges object with 2 ranges and 1 metadata column:
#>       seqnames    ranges strand |        revmap
#>          <Rle> <IRanges>  <Rle> | <IntegerList>
#>   [1]     chr2       4-9      - |             2
#>   [2]     chr1       1-3      - |             1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths
#>