Put (virtual) tiles on a given genome

tileGenome returns a set of genomic regions that form a partitioning of the specified genome. Each region is called a "tile".

tileGenome(seqlengths, ntile, tilewidth, cut.last.tile.in.chrom=FALSE)

Arguments

seqlengths

Either a named numeric vector of chromosome lengths or a Seqinfo object. More precisely, if a named numeric vector, it must have a length >= 1, cannot contain NAs or negative values, and cannot have duplicated names. If a Seqinfo object, then it's first replaced with the vector of sequence lengths stored in the object (extracted from the object with the seqlengths getter), then the restrictions described previously apply to this vector.

ntile

The number of tiles to generate.

tilewidth

The desired tile width. The effective tile width might be slightly different but is guaranteed to never be more than the desired width.

cut.last.tile.in.chrom

Whether or not to cut the last tile in each chromosome. This is set to FALSE by default. Can be set to TRUE only when tilewidth is specified. In that case, a tile will never overlap with more than 1 chromosome and a GRanges object is returned with one element (i.e. one genomic range) per tile.

Value

If cut.last.tile.in.chrom is FALSE (the default), a GRangesList object with one list element per tile, each of them containing a number of genomic ranges equal to the number of chromosomes it overlaps with. Note that when the tiles are small (i.e. much smaller than the chromosomes), most of them only overlap with a single chromosome.

If cut.last.tile.in.chrom is TRUE, a GRanges

object with one element (i.e. one genomic range) per tile.

Author

H. Pagès, based on a proposal by M. Morgan

See also

Examples

## ---------------------------------------------------------------------
## A. WITH A TOY GENOME
## ---------------------------------------------------------------------

seqlengths <- c(chr1=60, chr2=20, chr3=25)

## Create 5 tiles:
tiles <- tileGenome(seqlengths, ntile=5)
tiles
#> GRangesList object of length 5:
#> [[1]]
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1      1-21      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome
#> 
#> [[2]]
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1     22-42      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome
#> 
#> [[3]]
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1     43-60      *
#>   [2]     chr2       1-3      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome
#> 
#> [[4]]
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr2      4-20      *
#>   [2]     chr3       1-4      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome
#> 
#> [[5]]
#> GRanges object with 1 range and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr3      5-25      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome
#> 
elementNROWS(tiles)  # tiles 3 and 4 contain 2 ranges
#> [1] 1 1 2 2 1

width(tiles)
#> IntegerList of length 5
#> [[1]] 21
#> [[2]] 21
#> [[3]] 18 3
#> [[4]] 17 4
#> [[5]] 21
## Use sum() on this IntegerList object to get the effective tile
## widths:
sum(width(tiles))  # each tile covers exactly 21 genomic positions
#> [1] 21 21 21 21 21

## Create 9 tiles:
tiles <- tileGenome(seqlengths, ntile=9)
elementNROWS(tiles)  # tiles 6 and 7 contain 2 ranges
#> [1] 1 1 1 1 1 2 2 1 1

table(sum(width(tiles)))  # some tiles cover 12 genomic positions,
#> 
#> 11 12 
#>  3  6 
                          # others 11

## Specify the tile width:
tiles <- tileGenome(seqlengths, tilewidth=20)
length(tiles)  # 6 tiles
#> [1] 6
table(sum(width(tiles)))  # effective tile width is <= specified
#> 
#> 17 18 
#>  3  3 

## Specify the tile width and cut the last tile in each chromosome:
tiles <- tileGenome(seqlengths, tilewidth=24,
                    cut.last.tile.in.chrom=TRUE)
tiles
#> GRanges object with 6 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1      1-24      *
#>   [2]     chr1     25-48      *
#>   [3]     chr1     49-60      *
#>   [4]     chr2      1-20      *
#>   [5]     chr3      1-24      *
#>   [6]     chr3        25      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome
width(tiles)  # each tile covers exactly 24 genomic positions, except
#> [1] 24 24 12 20 24  1
              # the last tile in each chromosome

## Partition a genome by chromosome ("natural partitioning"):
tiles <- tileGenome(seqlengths, tilewidth=max(seqlengths),
                    cut.last.tile.in.chrom=TRUE)
tiles  # one tile per chromosome
#> GRanges object with 3 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chr1      1-60      *
#>   [2]     chr2      1-20      *
#>   [3]     chr3      1-25      *
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome

## sanity check
stopifnot(all.equal(setNames(end(tiles), seqnames(tiles)), seqlengths))

## ---------------------------------------------------------------------
## B. WITH A REAL GENOME
## ---------------------------------------------------------------------

library(BSgenome.Scerevisiae.UCSC.sacCer2)
tiles <- tileGenome(seqinfo(Scerevisiae), ntile=20)
tiles
#> GRangesList object of length 20:
#> [[1]]
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames    ranges strand
#>          <Rle> <IRanges>  <Rle>
#>   [1]     chrI  1-230208      *
#>   [2]    chrII  1-377942      *
#>   -------
#>   seqinfo: 18 sequences (2 circular) from sacCer2 genome
#> 
#> [[2]]
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames        ranges strand
#>          <Rle>     <IRanges>  <Rle>
#>   [1]    chrII 377943-813178      *
#>   [2]   chrIII      1-172914      *
#>   -------
#>   seqinfo: 18 sequences (2 circular) from sacCer2 genome
#> 
#> [[3]]
#> GRanges object with 2 ranges and 0 metadata columns:
#>       seqnames        ranges strand
#>          <Rle>     <IRanges>  <Rle>
#>   [1]   chrIII 172915-316617      *
#>   [2]    chrIV      1-464447      *
#>   -------
#>   seqinfo: 18 sequences (2 circular) from sacCer2 genome
#> 
#> ...
#> <17 more elements>

tiles <- tileGenome(seqinfo(Scerevisiae), tilewidth=50000,
                    cut.last.tile.in.chrom=TRUE)
tiles
#> GRanges object with 251 ranges and 0 metadata columns:
#>         seqnames        ranges strand
#>            <Rle>     <IRanges>  <Rle>
#>     [1]     chrI       1-50000      *
#>     [2]     chrI  50001-100000      *
#>     [3]     chrI 100001-150000      *
#>     [4]     chrI 150001-200000      *
#>     [5]     chrI 200001-230208      *
#>     ...      ...           ...    ...
#>   [247]   chrXVI 850001-900000      *
#>   [248]   chrXVI 900001-948062      *
#>   [249]     chrM       1-50000      *
#>   [250]     chrM   50001-85779      *
#>   [251]  2micron        1-6318      *
#>   -------
#>   seqinfo: 18 sequences (2 circular) from sacCer2 genome

## ---------------------------------------------------------------------
## C. AN APPLICATION: COMPUTE THE BINNED AVERAGE OF A NUMERICAL VARIABLE
##    DEFINED ALONG A GENOME
## ---------------------------------------------------------------------

## See '?genomicvars' for an example of how to compute the binned
## average of a numerical variable defined along a genome.