makeGRangesFromDataFrame {GenomicRanges}	R Documentation

Make a GRanges object from a data.frame or DataFrame

Description

makeGRangesFromDataFrame and makeGPosFromDataFrame both take a data-frame-like object as input and try to automatically find the columns that describe genomic ranges (for makeGRangesFromDataFrame) or genomic positions (for makeGPosFromDataFrame). If successful, they return them in a GRanges or GPos object.

The two functions are also the workhorses behind the coercion methods from data.frame (or DataFrame) to GRanges or GPos.

Usage

makeGRangesFromDataFrame(df, keep.extra.columns=FALSE,
        ignore.strand=FALSE, seqinfo=NULL,
        seqnames.field=c("seqnames", "seqname",
                         "chromosome", "chrom", "chr", "chromosome_name",
                         "seqid"),
        start.field="start",
        end.field=c("end", "stop"),
        strand.field="strand",
        starts.in.df.are.0based=FALSE,
        na.rm=FALSE)

makeGPosFromDataFrame(df, keep.extra.columns=FALSE,
        ignore.strand=FALSE, seqinfo=NULL,
        seqnames.field=c("seqnames", "seqname",
                         "chromosome", "chrom", "chr", "chromosome_name",
                         "seqid"),
        pos.field=c("pos", "position", "positions"),
        strand.field="strand",
        na.rm=FALSE)

Arguments

df
A data.frame or DataFrame object. If not, then the function first tries to turn df into a data frame with as.data.frame(df).

keep.extra.columns
TRUE or FALSE (the default). If TRUE, then the columns in df that are not used to form the genomic ranges or genomic positions of the returned GRanges or GPos object are returned as metadata columns on the object. Otherwise, they are ignored.

Note that if df has a width column, then makeGRangesFromDataFrame will always ignore it.

ignore.strand
TRUE or FALSE (the default). If TRUE, then the strand of the returned GRanges or GPos object is set to "*".

seqinfo
Either NULL, or a Seqinfo object, or a character vector of unique sequence names (a.k.a. seqlevels), or a named numeric vector of sequence lengths. When not NULL, seqinfo must be compatible with the genomic ranges in df, that is, it must have one entry for each unique sequence name represented in df. Note that it can have additional entries i.e. entries for seqlevels not represented in df.

seqnames.field
A character vector of recognized names for the column in df that contains the chromosome name (a.k.a. sequence name) associated with each genomic range. Only the first name in seqnames.field that is found in colnames(df) is used. If no one is found, then an error is raised.

start.field
A character vector of recognized names for the column in df that contains the start positions of the genomic ranges. Only the first name in start.field that is found in colnames(df) is used. If no one is found, then an error is raised.

end.field
A character vector of recognized names for the column in df that contains the end positions of the genomic ranges. Only the first name in start.field that is found in colnames(df) is used. If no one is found, then an error is raised.

pos.field
A character vector of recognized names for the column in df that contains the genomic positions to use to make the GPos object. Only the first name in pos.field that is found in colnames(df) is used. If no one is found, then an error is raised.

strand.field
A character vector of recognized names for the column in df that contains the strand associated with each genomic range. Only the first name in strand.field that is found in colnames(df) is used. If no one is found or if ignore.strand is TRUE, then the strand of the returned GRanges object is set to "*".

starts.in.df.are.0based
TRUE or FALSE (the default). If TRUE, then the start positions of the genomic ranges in df are considered to be 0-based and are converted to 1-based in the returned GRanges object. This feature is intended to make it more convenient to handle input that contains data obtained from resources using the "0-based start" convention. A notorious example of such resource is the UCSC Table Browser (http://genome.ucsc.edu/cgi-bin/hgTables).

na.rm
TRUE or FALSE (the default).

If TRUE, then rows in the df with missing start, end, or position values (i.e. the value is NA) are ignored. Otherwise, they trigger an error.

Value

A GRanges or GPos object.

If na.rm is set to FALSE (the default), then the returned object is guaranteed to have one element per row in the input. However, if na.rm is set to TRUE, then the length of the returned object can be less than nrow(df).

If the seqinfo argument was supplied, the returned object will have exactly the seqlevels specified in seqinfo and in the same order. Otherwise, the seqlevels are ordered according to the output of the rankSeqlevels function (except if df contains the seqnames in the form of a factor or factor-Rle, in which case the levels() of the seqnames become the seqlevels of the returned object as-is, that is, without any re-ordering).

If df has non-automatic row names (i.e. rownames(df) is not NULL and is not seq_len(nrow(df))), then they will be used to set names on the returned GRanges or GPos object.

Note

Coercing a data.frame or DataFrame df to GRanges (with as(df, "GRanges")), or calling GRanges(df), are both equivalent to calling makeGRangesFromDataFrame(df, keep.extra.columns=TRUE).

Coercing a data.frame or DataFrame df to GPos (with as(df, "GPos")) is equivalent to calling makeGPosFromDataFrame(df, keep.extra.columns=TRUE).

Author(s)

H. Pagès, based on a proposal by Kasper Daniel Hansen

See Also

GRanges and GPos objects.

Seqinfo objects and the rankSeqlevels function in the Seqinfo package.

The makeGRangesListFromFeatureFragments function for making a GRangesList object from a list of fragmented features.

makeIRangesFromDataFrame in the IRanges package to make an IRanges object from a data.frame or DataFrame.

DataFrame objects in the S4Vectors package.

Examples

## ---------------------------------------------------------------------
## BASIC makeGRangesFromDataFrame() EXAMPLES
## ---------------------------------------------------------------------

df <- data.frame(chr="chr1", start=11:15, end=12:16,
                 strand=c("+","-","+","*","."), score=1:5)
df
makeGRangesFromDataFrame(df)  # strand value "." is replaced with "*"

## With NAs in range-defining columns:
df$start[5] <- df$end[2] <- NA
df
#makeGRangesFromDataFrame(df)  # error!
makeGRangesFromDataFrame(df, na.rm=TRUE)  # rows with NAs got dropped

## The strand column is optional:
df <- data.frame(chr="chr1", start=11:15, end=12:16, score=1:5)
makeGRangesFromDataFrame(df)

gr <- makeGRangesFromDataFrame(df, keep.extra.columns=TRUE)
gr2 <- as(df, "GRanges")  # equivalent to the above
stopifnot(identical(gr, gr2))
gr2 <- GRanges(df)        # equivalent to the above
stopifnot(identical(gr, gr2))

makeGRangesFromDataFrame(df, ignore.strand=TRUE)
makeGRangesFromDataFrame(df, keep.extra.columns=TRUE,
                             ignore.strand=TRUE)

makeGRangesFromDataFrame(df, seqinfo=paste0("chr", 4:1))
makeGRangesFromDataFrame(df, seqinfo=c(chrM=NA, chr1=500, chrX=100))
makeGRangesFromDataFrame(df, seqinfo=Seqinfo(paste0("chr", 4:1)))

## ---------------------------------------------------------------------
## BASIC makeGPosFromDataFrame() EXAMPLES
## ---------------------------------------------------------------------

df <- data.frame(
    rsids=    c("rs7927381", "rs79273813", "rs79273815",
                "rs79273819", "rs79273817", "rs79273811", "rs79273814"),
    seqnames= c("11", "1", "11", "1", "5", "8", "8"),
    positions=c("67579271", "179599437", "70746832",
                "5066529", "107006951", "95077418", "43582071"),
    alleles=  c("T/C", "A/G", "C/G, C/T", "G/A, G/C", "T/C, T/G",
                "G/A, G/T", "G/A, G/T"),
    genes=    c(NA, "TDRD5", "SHANK2", NA, "LOC102467213", "NDUFAF6", NA),
    row.names=letters[1:7]
)

makeGPosFromDataFrame(df)
makeGPosFromDataFrame(df, keep.extra.columns=TRUE)
as(df, "GPos")  # same as above

## With missing positions:
df$positions[2L] <- NA
df
#makeGPosFromDataFrame(df)  # error!
makeGPosFromDataFrame(df, na.rm=TRUE)  # rows with NAs got dropped

## ---------------------------------------------------------------------
## ABOUT AUTOMATIC DETECTION OF THE seqnames/start/end/strand COLUMNS
## ---------------------------------------------------------------------

## Automatic detection of the seqnames/start/end/strand columns is
## case insensitive:
df <- data.frame(ChRoM="chr1", StarT=11:15, stoP=12:16,
                 STRAND=c("+","-","+","*","."), score=1:5)
makeGRangesFromDataFrame(df)

## It also ignores a common prefix between the start and end columns:
df <- data.frame(seqnames="chr1", tx_start=11:15, tx_end=12:16,
                 strand=c("+","-","+","*","."), score=1:5)
makeGRangesFromDataFrame(df)

## The common prefix between the start and end columns is used to
## disambiguate between more than one seqnames column:
df <- data.frame(chrom="chr1", tx_start=11:15, tx_end=12:16,
                 tx_chr="chr2", score=1:5)
makeGRangesFromDataFrame(df)

## ---------------------------------------------------------------------
## 0-BASED VS 1-BASED START POSITIONS
## ---------------------------------------------------------------------

if (require(UCSC.utils)) {
  ## Use list_UCSC_tracks() to lookup the primary table associated with
  ## a given track:
  subset(list_UCSC_tracks("sacCer2"), track=="Assembly")

  df <- fetch_UCSC_track_data("sacCer2", primary_table="gold")
  head(df)

  ## A common pitfall is to forget that UCSC uses the "0-based start"
  ## convention to store genomic ranges in their databases:
  gr0 <- makeGRangesFromDataFrame(df, keep.extra.columns=TRUE,
                                      start.field="chromStart",
                                      end.field="chromEnd")
  head(gr0)

  ## The start positions need to be converted into 1-based positions,
  ## to adhere to the convention used in Bioconductor:
  gr1 <- makeGRangesFromDataFrame(df, keep.extra.columns=TRUE,
                                      start.field="chromStart",
                                      end.field="chromEnd",
                                      starts.in.df.are.0based=TRUE)
  head(gr1)

  ## IMPORTANT NOTE: The start positions in the "fragStart" metadata
  ## column are still 0-based! That's because using
  ## 'keep.extra.columns=TRUE' only takes care of converting the
  ## start positions of the genomic ranges that define the GRanges
  ## object itself, but it doesn't meddle with its metadata columns.
}

[Package GenomicRanges version 1.62.1 Index]